ai_services4/resume-analyzer/requirements.txt
# ============================================
# Resume Analyzer Service - Requirements
# Port: 8000
# ============================================

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.2
starlette==0.27.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.0
pypdfium2==4.30.0
pdfminer.six==20221105

# NLP & ML
nltk==3.8.1
spacy==3.7.2
sentence-transformers==2.2.2
transformers==4.33.2
torch==2.2.0
numpy==1.24.3
scikit-learn==1.3.0

# Data Processing
pandas==2.0.3

# PDF Generation
fpdf2==2.8.4

# Google APIs
google-generativeai==0.8.5
google-api-python-client==2.182.0
google-auth==2.40.3

# Security
python-jose==3.5.0
passlib==1.7.4

# Database
pymongo==4.14.1

# Utilities
requests==2.31.0
beautifulsoup4==4.13.5

# Audio (if needed)
librosa==0.10.1
soundfile==0.12.1

# Additional dependencies from main
rsa==4.9.1
safetensors==0.6.2
scipy==1.15.3
sentencepiece==0.2.1
setuptools==65.5.0
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.8
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
thinc==8.2.5
threadpoolctl==3.6.0
tokenizers==0.13.3
torchvision==0.23.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.9.4
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wasabi==1.1.3
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.3.4
websockets==15.0.1

# Spacy model
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# Note: Install spacy model separately after requirements:
# python -m spacy download en_core_web_sm
